Convert v1 character sheet to run from v2 character sheet fix statuses removal, add warning message about upcoming sheet removal
