+++ styles/system/main.less
@@ -480,18 +480,29 @@
       cursor: default;
     }
   }
-  .sheet-tabs {
+  .sheet-nav {
     flex: 0 0 20px;
-    margin-bottom: 5px;
+    margin: 4px 0 5px;
+    line-height: 20px;
     font-family: customSheetFont, 'Palatino Linotype', serif;
     font-size: 20px;
     border: 0;
-    .item {
+    .tab-name {
       line-height: 15px;
       margin: 0 10px;
       border-bottom: 3px solid #b5b3a4;
+      text-align: center;
+      &.keeper-only-tab {
+        flex: 0.3;
+        span {
+          vertical-align: bottom;
+          i {
+            color: @colorGreen;
+          }
+        }
+      }
     }
-    .item.active {
+    .tab-name.active {
       border-bottom: 3px solid #44191a;
     }
   }
